fix(app): ajuste em tipo de dado de preco

+++ src/main/java/com/project/extension/entity/Produto.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -22,7 +23,7 @@ public class Produto extends BaseEntity {
     @Column(name = "unidade_medida")
     private String unidademedida;
 
-    private Double preco;
+    private BigDecimal preco;
     private Boolean ativo;
 
     @OneToMany(mappedBy = "produto")
@@ -34,7 +35,7 @@ public class Produto extends BaseEntity {
     @OneToOne(cascade = CascadeType.ALL)
     private MetricaEstoque metricaEstoque;
 
-    public Produto(String nome, String descricao, String unidademedida, Double preco, Boolean ativo) {
+    public Produto(String nome, String descricao, String unidademedida, BigDecimal preco, Boolean ativo) {
         this.nome = nome;
         this.descricao = descricao;
         this.unidademedida = unidademedida;
